@@ -9,15 +9,15 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "16"
   },
   "main": "lib/index.js",
   "dependencies": {
-    "firebase-admin": "^9.2.0",
-    "firebase-functions": "^3.11.0"
+    "firebase-admin": "^10.0.2",
+    "firebase-functions": "^3.18.0"
   },
   "devDependencies": {
-    "typescript": "^3.8.0",
+    "typescript": "^4.5.4",
     "firebase-functions-test": "^0.2.0"
   },
   "private": true
