@@ -1,10 +1,9 @@
-import {Injectable} from "@angular/core";
-import {Observable} from "rxjs";
-import {AngularFireAuth} from "@angular/fire/auth";
-import {map} from "rxjs/operators";
-import {Router} from "@angular/router";
-import {UserRoles} from "../model/user-roles";
-
+import { Injectable } from "@angular/core";
+import { AngularFireAuth } from "@angular/fire/auth";
+import { Observable } from "rxjs";
+import { map } from "rxjs/operators";
+import { Router } from '@angular/router';
+import { UserRoles } from "../model/user-roles"
 
 @Injectable({
     providedIn: "root"
@@ -12,31 +11,21 @@ import {UserRoles} from "../model/user-roles";
 export class UserService {
 
     isLoggedIn$ : Observable<boolean>;
-
     isLoggedOut$: Observable<boolean>;
-
-    pictureUrl$: Observable<string>;
-
-    roles$ : Observable<UserRoles>;
+    pictureUrl$ : Observable<string>;
+    roles$      : Observable<UserRoles>;
 
     constructor(
-        private afAuth: AngularFireAuth,
-        private router: Router) {
-
+            private afAuth: AngularFireAuth,
+            private router: Router) {
         this.isLoggedIn$ = afAuth.authState.pipe(map(user => !!user));
-
         this.isLoggedOut$ = this.isLoggedIn$.pipe(map(loggedIn => !loggedIn));
-
-        this.pictureUrl$ =
-            afAuth.authState.pipe(map(user => user? user.photoURL : null));
-
+        this.pictureUrl$ = afAuth.authState.pipe(map(user => user? user.photoURL : null));
         this.roles$ = this.afAuth.idTokenResult
             .pipe(
                 map(token => <any>token?.claims ?? {admin:false})
             )
-
     }
-
 
     logout() {
         this.afAuth.signOut();
