@@ -1,37 +1,27 @@
-
 import * as functions from 'firebase-functions';
-import {auth} from "./init";
-
-
+import { auth } from './init';
 
 export function getUserCredentialsMiddleware(req, res, next) {
 
-    functions.logger.debug(`Attempting to extract user credentials from request.`);
+    functions.logger.debug(`Attemping to extract user credentials from request`);
 
     const jwt = req.headers.authorization;
 
-
-    if (jwt) {
+    if(jwt) {
         auth.verifyIdToken(jwt)
             .then(jwtPayload => {
+                req["uid"] = jwtPayload.uid;
+                req["admin"] = jwtPayload.admin;
 
-                 req["uid"] = jwtPayload.uid;
-                 req["admin"] = jwtPayload.admin;
-
-                functions.logger.debug(
-                    `Credentials: uid=${jwtPayload.uid}, admin=${jwtPayload.admin}`);
-
+                functions.logger.debug(`Credentials: uid=${jwtPayload.uid}, admin=${jwtPayload.admin}`);
                 next();
             })
             .catch(err => {
-                console.log("Error ocurred while validating JWT", err);
+                console.log("Error while validating JWT", err);
                 next();
             });
-
     }
     else {
         next();
     }
-
-
 }
