@@ -1,25 +1,24 @@
 import * as functions from "firebase-functions";
-
 import {db} from "../init";
+import {firestore} from 'firebase-admin';
 
-import {firestore} from 'firebase-admin/lib/firestore';
-import FieldValue = firestore.FieldValue;
+// // Start writing Firebase Functions
+// // https://firebase.google.com/docs/functions/typescript
 
-export default async(change, context) => {
+export default async (change, context) => {
 
     if(context.params.courseId == 'stats') {
         return;
     }
 
-    functions.logger.debug(
-        `Running update course trigger for courseId ${context.params.courseId}`);
+    functions.logger.debug(`Running add course trigger ${context.params.courseId}`);
 
     const newData = change.after.data(),
             oldData = change.before.data();
 
     let increment = 0;
 
-    if (!oldData.promo && newData.promo) {
+    if(!oldData.promo && newData.promo) {
         increment = 1;
     }
     else if (oldData.promo && !newData.promo) {
@@ -31,7 +30,6 @@ export default async(change, context) => {
     }
 
     return db.doc(`courses/stats`).update({
-        totalPromo: FieldValue.increment(increment)
+        totalPromo: firestore.FieldValue.increment(increment)
     })
-
 }
