@@ -4,7 +4,7 @@ import {Course} from '../model/course';
 import {finalize, tap} from 'rxjs/operators';
 import {Observable} from 'rxjs';
 import {Lesson} from '../model/lesson';
-import {CoursesService} from "../services/courses.service";
+import { CoursesService } from '../services/courses.service';
 
 
 @Component({
@@ -14,7 +14,7 @@ import {CoursesService} from "../services/courses.service";
 })
 export class CourseComponent implements OnInit {
 
-  course:Course;
+  course: Course;
 
   lessons: Lesson[];
 
@@ -25,41 +25,37 @@ export class CourseComponent implements OnInit {
   displayedColumns = ['seqNo', 'description', 'duration'];
 
   constructor(
-      private route: ActivatedRoute,
-      private coursesService: CoursesService) {
+    private route: ActivatedRoute,
+    private coursesService: CoursesService) {
 
   }
 
   ngOnInit() {
+    this.course = this.route.snapshot.data["course"];
 
-      this.course = this.route.snapshot.data["course"];
+    this.loading = true;
 
-      this.loading = true;
-
-      this.coursesService.findLessons(this.course.id)
-          .pipe(
-              finalize(() => this.loading = false)
-          )
-          .subscribe(
-              lessons => this.lessons = lessons
-          );
-
+    this.coursesService.findLessons(this.course.id)
+      .pipe(
+        finalize(() => this.loading = false)
+      )
+      .subscribe(
+        lessons => this.lessons = lessons
+      )
   }
 
-    loadMore() {
+  loadMore() {
+    this.lastPageLoaded++;
 
-      this.lastPageLoaded++;
+    this.loading = true;
 
-      this.loading = true;
-
-      this.coursesService.findLessons(this.course.id, "asc",
-          this.lastPageLoaded)
-          .pipe(
-              finalize(() => this.loading = false)
-          )
-          .subscribe(lessons => this.lessons = this.lessons.concat(lessons))
-
-    }
+    this.coursesService.findLessons(this.course.id, "asc",
+      this.lastPageLoaded)
+      .pipe(
+        finalize(() => this.loading = false)
+      )
+      .subscribe(lessons => this.lessons = this.lessons.concat(lessons))
+  }
 }
 
 
