@@ -5,8 +5,8 @@ import {EditCourseDialogComponent} from "../edit-course-dialog/edit-course-dialo
 import {catchError, tap} from 'rxjs/operators';
 import {throwError} from 'rxjs';
 import {Router} from '@angular/router';
-import {CoursesService} from "../services/courses.service";
-import {UserService} from "../services/user.service";
+import { CoursesService } from '../services/courses.service';
+import { UserService } from '../services/user.service';
 
 @Component({
     selector: 'courses-card-list',
@@ -70,10 +70,7 @@ export class CoursesCardListComponent implements OnInit {
                 })
             )
             .subscribe();
-
     }
-
-
 }
 
 
