@@ -4,7 +4,7 @@ import {Course} from "../model/course";
 import {FormBuilder, Validators, FormGroup} from "@angular/forms";
 import {AngularFireStorage} from '@angular/fire/storage';
 import {Observable} from 'rxjs';
-import {CoursesService} from "../services/courses.service";
+import { CoursesService } from '../services/courses.service';
 
 
 @Component({
@@ -14,7 +14,7 @@ import {CoursesService} from "../services/courses.service";
 })
 export class EditCourseDialogComponent {
 
-    form:FormGroup;
+    form: FormGroup;
 
     course: Course;
 
@@ -22,41 +22,26 @@ export class EditCourseDialogComponent {
         private dialogRef: MatDialogRef<EditCourseDialogComponent>,
         private fb: FormBuilder,
         @Inject(MAT_DIALOG_DATA) course: Course,
-        private coursesService: CoursesService
+        private coursesService: CoursesService,
     ) {
-
         this.course = course;
-
+        
         this.form = this.fb.group({
             description: [course.description, Validators.required],
             longDescription: [course.longDescription, Validators.required],
             promo: [course.promo]
-        });
-
+        })
     }
 
     close() {
         this.dialogRef.close();
-
     }
 
     save() {
-
         const changes = this.form.value;
-
         this.coursesService.updateCourse(this.course.id, changes)
             .subscribe(() => {
-
                 this.dialogRef.close(changes);
-
             });
-
-
     }
 }
-
-
-
-
-
-
