@@ -1,18 +1,15 @@
-import {Injectable} from "@angular/core";
-import {AngularFireAuth} from "@angular/fire/auth";
-
+import { Injectable } from "@angular/core";
+import { AngularFireAuth } from "@angular/fire/auth";
 
 @Injectable({
     providedIn: "root"
 })
 export class AuthTokenService {
-
-    authJwtToken:string;
+    
+    authJwtToken: string;
 
     constructor(private afAuth: AngularFireAuth) {
         afAuth.idToken
             .subscribe(jwt => this.authJwtToken = jwt);
     }
-
-
 }
