@@ -1,21 +1,17 @@
-import {Injectable} from "@angular/core";
-import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from "@angular/common/http";
-import {Observable} from "rxjs";
-import {AuthTokenService} from "./auth-token.service";
-
+import { Injectable } from "@angular/core";
+import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from "@angular/common/http";
+import { Observable } from "rxjs";
+import { AuthTokenService } from "./auth-token.service";
 
 @Injectable()
-export class AuthInterceptor implements  HttpInterceptor {
+export class AuthInterceptor implements HttpInterceptor {
 
-    constructor(private token: AuthTokenService) {
-
+    constructor(private token: AuthTokenService) {   
     }
-
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
 
-        if (this.token.authJwtToken) {
-
+        if(this.token.authJwtToken) {
             const cloned = req.clone({
                 headers: req.headers
                     .set("Authorization", this.token.authJwtToken)
@@ -28,5 +24,4 @@ export class AuthInterceptor implements  HttpInterceptor {
         }
 
     }
-
 }
