@@ -1,11 +1,8 @@
-
-
-export function convertSnaps<T>(results)  {
-    return <T[]> results.docs.map(snap => {
+export function convertSnaps<T>(results) {
+    return <T[]>results.docs.map(snap => {
         return {
             id: snap.id,
             ...<any>snap.data()
         }
     })
-
 }
