@@ -6,15 +6,16 @@ export const environment = {
   production: false,
   useEmulators: true,
   firebase: {
-      apiKey: "AIzaSyB9LOREMGhj1jpVXOHTKIwQu2oM7pVfjQg",
-      authDomain: "fir-course-recording-c7f3e.firebaseapp.com",
-      projectId: "fir-course-recording-c7f3e",
-      storageBucket: "fir-course-recording-c7f3e.appspot.com",
-      messagingSenderId: "927953565493",
-      appId: "1:927953565493:web:0d4a8e79cc45fd38733e7c"
+    apiKey: "AIzaSyBYAoG5NaTpvjZS1mcYMnm4GzEaXdwyolc",
+    authDomain: "fortun-8.firebaseapp.com",
+    projectId: "fortun-8",
+    storageBucket: "fortun-8.appspot.com",
+    messagingSenderId: "789104168940",
+    appId: "1:789104168940:web:45c7b91a9d39cdd439fac9",
+    measurementId: "G-VH3EC06HGM"  
   },
   api: {
-      createUser: "http://localhost:5001/fir-course-recording-c7f3e/us-central1/createUser"
+    createUser: "http://localhost:5001/fortun-8/us-central1/createUser"
   }
 };
 
